ForecastActivity --> ForecastFragment (+few corrections in the fragments)
